pyproject.toml
[tool.poetry]
name = "mixtape"
version = "0.1.0"
description = ""
authors = ["Sebastian Höffner <info@sebastian-hoeffner.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-mpd2 = "^1.0.0"
python-telegram-bot = "^12.4.2"
youtube-dl = "^2020.2.16"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
